# DevShop/openapi_mcp.json
{"openapi":"3.0.3","info":{"title":"devShop Product Reviews API","description":"REST API for managing product reviews and ratings","version":"v1","contact":{"name":"devShop API Team","email":"api@devshop.com"},"license":{"name":"MIT License","url":"https://opensource.org/licenses/MIT"}},"tags":[{"name":"ProductInventory","description":"Product inventory management endpoints"},{"name":"ProductReviews","description":"Product review management endpoints"}],"paths":{"/api/inventory":{"get":{"tags":["ProductInventory"],"summary":"Get all products in inventory with stock levels","description":"Retrieves all products with inventory information","operationId":"GetAllInventory","responses":{"200":{"description":"Successfully retrieved inventory","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductInventoryItem"}}},"application/xml":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductInventoryItem"}}}}},"500":{"description":"Internal server error occurred","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}}}}}},"/api/inventory/product/{item_id}":{"get":{"tags":["ProductInventory"],"summary":"Get inventory details for a specific product","description":"Retrieves inventory information for a single product","operationId":"GetProductInventory","parameters":[{"name":"item_id","in":"path","description":"The unique identifier of the product","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Successfully retrieved product inventory","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductInventoryItem"}},"application/xml":{"schema":{"$ref":"#/components/schemas/ProductInventoryItem"}}}},"404":{"description":"Product not found","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}}},"500":{"description":"Internal server error occurred","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}}}}}},"/api/inventory/category/{categoryName}":{"get":{"tags":["ProductInventory"],"summary":"Get all products in a specific category with inventory","description":"Retrieves inventory for products in the specified category","operationId":"GetInventoryByCategory","parameters":[{"name":"categoryName","in":"path","description":"The name of the category","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully retrieved inventory for category","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductInventoryItem"}}}}},"400":{"description":"Category name is required","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}}},"500":{"description":"Internal server error occurred","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}}}}}},"/api/inventory/product/{productId}/size/{size}":{"get":{"tags":["ProductInventory"],"summary":"Check stock availability for a specific product and size","description":"Checks if a specific size is available for the product","operationId":"CheckStockAvailability","parameters":[{"name":"productId","in":"path","description":"The unique identifier of the product","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"size","in":"path","description":"Size code (XS, S, M, L, XL, XXL, XXXL)","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully retrieved stock availability","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StockAvailability"}}}},"400":{"description":"Invalid size specified","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}}},"404":{"description":"Product not found","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}}},"500":{"description":"Internal server error occurred","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}}}}}},"/api/inventory/lowstock":{"get":{"tags":["ProductInventory"],"summary":"Get products with low stock","description":"Retrieves products below the stock threshold","operationId":"GetLowStockProducts","parameters":[{"name":"threshold","in":"query","description":"Stock quantity threshold (default: 5)","required":false,"schema":{"type":"integer","format":"int32","default":5}}],"responses":{"200":{"description":"Successfully retrieved low stock products","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LowStockItem"}}}}},"500":{"description":"Internal server error occurred","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}}}}}},"/api/products/{productId}/reviews":{"get":{"tags":["ProductReviews"],"summary":"Get all reviews for a specific product","description":"Retrieves all reviews for the specified product","operationId":"GetProductReviews","parameters":[{"name":"productId","in":"path","description":"The unique identifier of the product","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Successfully retrieved product reviews","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductReview"}}}}},"500":{"description":"Internal server error occurred","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}}}}}},"/api/reviews":{"get":{"tags":["ProductReviews"],"summary":"Get all product reviews across all products","description":"Retrieves all product reviews in the system","operationId":"GetAllReviews","responses":{"200":{"description":"Successfully retrieved all reviews","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductReview"}}}}},"500":{"description":"Internal server error occurred","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}}}}}},"/api/reviews/category/{categoryName}":{"get":{"tags":["ProductReviews"],"summary":"Get all reviews filtered by category name","description":"Retrieves reviews for products in the specified category","operationId":"GetReviewsByCategory","parameters":[{"name":"categoryName","in":"path","description":"The name of the category","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully retrieved reviews for the category","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductReview"}}}}},"400":{"description":"Category name is required","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}}},"500":{"description":"Internal server error occurred","headers":{"X-Request-ID":{"description":"Unique request identifier for tracking","schema":{"type":"string"}}}}}}}},"components":{"schemas":{"ProductInventoryItem":{"type":"object","description":"Product inventory item with stock details","properties":{"id":{"type":"integer","format":"int32","description":"Product identifier"},"name":{"type":"string","nullable":true,"description":"Product name"},"category":{"type":"string","nullable":true,"description":"Product category"},"price":{"type":"number","format":"double","description":"Product price"},"description":{"type":"string","nullable":true,"description":"Product description"},"sizes":{"type":"object","description":"Stock quantities by size","additionalProperties":{"type":"integer","format":"int32"}},"totalStock":{"type":"integer","format":"int32","description":"Total stock across all sizes"}}},"StockAvailability":{"type":"object","description":"Stock availability for a specific product and size","properties":{"productId":{"type":"integer","format":"int32","description":"Product identifier"},"productName":{"type":"string","nullable":true,"description":"Product name"},"size":{"type":"string","nullable":true,"description":"Size code"},"quantity":{"type":"integer","format":"int32","description":"Available quantity"},"isInStock":{"type":"boolean","description":"Whether the item is in stock"},"status":{"type":"string","nullable":true,"description":"Stock status message"}}},"LowStockItem":{"type":"object","description":"Low stock item details","properties":{"productId":{"type":"integer","format":"int32","description":"Product identifier"},"productName":{"type":"string","nullable":true,"description":"Product name"},"category":{"type":"string","nullable":true,"description":"Category name"},"size":{"type":"string","nullable":true,"description":"Size with low stock"},"quantity":{"type":"integer","format":"int32","description":"Current quantity"},"alertLevel":{"type":"string","nullable":true,"description":"Alert level (Critical, Warning, Low)"}}},"ProductReview":{"type":"object","description":"Product review model","properties":{"reviewId":{"type":"integer","format":"int32","description":"Review identifier"},"productId":{"type":"integer","format":"int32","description":"Product identifier"},"reviewerName":{"type":"string","nullable":true,"description":"Name of the reviewer"},"rating":{"type":"integer","format":"int32","minimum":1,"maximum":5,"description":"Rating from 1 to 5 stars"},"comments":{"type":"string","nullable":true,"description":"Review comments"},"reviewDate":{"type":"string","format":"date-time","description":"Review date"},"product":{"type":"string","nullable":true,"description":"Product name"},"category":{"type":"string","nullable":true,"description":"Category name"}}}}}}